+++ pkg/godax/fills_test.go
@@ -1,64 +1,9 @@
 package godax
 
 import (
-	"net/http"
 	"reflect"
 	"testing"
 )
-
-const (
-	baseRestURL = "https://test.pro.coinbase"
-	baseWsURL   = "wss://test.ws-feed.pro.coinbase.com"
-	key         = "super_secret_key_123_abc"
-	secret      = "MTIzYWJjU3VwZXJTZWNyZXRTZWNyZXQ="
-	passphrase  = "1q2w3e4r"
-)
-
-type fields struct {
-	baseRestURL string
-	baseWsURL   string
-	key         string
-	secret      string
-	passphrase  string
-	httpClient  *http.Client
-}
-
-func defaultFields() fields {
-	return fields{
-		baseRestURL: baseRestURL,
-		baseWsURL:   baseWsURL,
-		key:         key,
-		secret:      secret,
-		passphrase:  passphrase,
-	}
-}
-
-func validateHeaders(t *testing.T, client *Client) {
-	compareHeader(t, client, "CB-ACCESS-KEY", key)
-	compareHeader(t, client, "CB-ACCESS-PASSPHRASE", passphrase)
-	compareHeader(t, client, "User-Agent", userAgent)
-	compareHeader(t, client, "Content-Type", "application/json")
-	compareHeader(t, client, "Accept", "application/json")
-	validateHeaderPresent(t, client, "CB-ACCESS-SIGN")
-	validateHeaderPresent(t, client, "CB-ACCESS-TIMESTAMP")
-}
-
-func compareHeader(t *testing.T, c *Client, wantHeader string, wantContent string) {
-	if c.httpClient.(*MockClient).Requests[0].Header.Get(wantHeader) != wantContent {
-		t.Errorf(
-			"%s header should be %s, was '%s'\n",
-			wantHeader,
-			wantContent,
-			c.httpClient.(*MockClient).Requests[0].Header.Get(wantHeader),
-		)
-	}
-}
-
-func validateHeaderPresent(t *testing.T, c *Client, wantHeader string) {
-	if c.httpClient.(*MockClient).Requests[0].Header.Get(wantHeader) == "" {
-		t.Errorf("%s header should not be empty\n", wantHeader)
-	}
-}
 
 // TODO: figure file
 func TestClient_ListFills(t *testing.T) {
